Add value comparer to string list property
Using test :CanCreateAPawtionContextUsingSqlite
Before adding the ".Metadata.SetValueComparer" in the dbContext mapping the resulting
list in the db will not be updated when calling SaveChanges the second time.
This is because the ChangeTracker will compare this list by reference .
Adding an explicit value comparer forces EF to compare the actual values in the list
so changes to the original list will be picked up and the db will be updated.